Move winamp sdk files to subfolder for clarity
