+++ winamp/sdk/ipc_pe.h
