+++ winamp/sdk/wa_ipc.h
